webapp/code/sriptRatingPrediction/requirements.txt
pandas
numpy
streamlit
PIL
wordcloud
collections
nltk
re
pickle
keras-preprocessing
keras.models